ci: the 386 matrix only compiled, and two 32-bit defects were sitting behind it
cross-build proves every module COMPILES on linux/386. Compiling and behaving
are different questions, and the gap is where a 64-bit assumption survives:
int(0xffffffff) is -1 where int is 32 bits and passes every "> cap" check —
exactly the bound internal/service/session's at-rest frame relies on, whose two
malformed-frame cases can only fail there.

Running the four workspace modules on linux/386 found two defects at once. A
bench helper did not compile, because Statfs_t.Type is int32 on 386 and int64
elsewhere; both the switch and the log now compare at one width. And
encwrite's frame-overflow test synthesises a slice longer than a 32-bit len can
hold — "unsafe.Slice: len out of range" — for a guard that cannot fire there at
all, since len never reaches MaxUint32 on a 32-bit word; it now skips, naming
the reason.

The job needs no emulation (amd64 runners execute 386 binaries natively) and
runs without -race, which has no 386 support — which also makes it the second
lane to compile the //go:build !race files.

The workflows' "no go test lane" rule is amended rather than contradicted: what
it forbids is a second lane for a platform Bazel already covers. Bazel here
builds for the host, so for 386 the alternative to raw go is no coverage.

+++ internal/service/writer/file/file_linux_bench_test.go
@@ -102,7 +102,9 @@ func fsName(b *testing.B, dir string) string {
 		//: an unreadable filesystem type is stated, never guessed.
 		return "unknown"
 	}
-	switch st.Type {
+	//: Statfs_t.Type is int64 on 64-bit Linux and int32 on 386, so the
+	//: comparison is made at one width rather than at the platform's.
+	switch int64(st.Type) {
 	//: the page cache, no device behind it.
 	case magicTmpfs:
 		return "tmpfs"
@@ -118,7 +120,7 @@ func fsName(b *testing.B, dir string) string {
 	//: anything else is visibly unrecognised, and the magic goes to the log
 	//: rather than into a row name nobody can compare across machines.
 	default:
-		b.Logf("statfs %s: unrecognised magic %#x", dir, st.Type)
+		b.Logf("statfs %s: unrecognised magic %#x", dir, int64(st.Type))
 		return "unrecognised"
 	}
 }
